#116: Move example contract into template package and add Compiler class for experimentation.
